\name{testarma1-package}
\alias{testarma1-package}
\alias{testarma1}
\docType{package}
\title{
\packageTitle{testarma1}
}
\description{
\packageDescription{testarma1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testarma1}
\packageIndices{testarma1}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{testarma1}

Maintainer: \packageMaintainer{testarma1}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
